refactor(gateway): 新增 anthropic passthrough TK companion
请求预处理、400 rectifier 与 stream terminal 迁入 gateway_anthropic_passthrough_tk.go。
功能不变。no-web-impact
